pyproject.toml
[project]
name = "pdebenchmark"
description = "pdebenchmark is tool for generating pde benchmark solutions"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "absl-py",
  "jaxlib",
  "jax",
  "numpy",
  "h5py",
  "diffrax",
]

dynamic = ["version"]


[project.optional-dependencies]
dev = [
  "pylint>=2.6.0",
  "yapf"
]

[tool.setuptools]
py-modules = ["pdebenchmark"]
